requirements.txt
python-dotenv
pymongo
streamlit
altair
pandas